pyproject.toml
[project]
name = "p2"
version = "0.1.0"
description = ""
authors = [
    {name = "Vladyslav",email = "vladkolodka@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psutil (>=7.0.0,<8.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "python-telegram-bot (>=22.0,<23.0)",
    "jupyter (>=1.1.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "p2", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
